remove pythonpath definition
