+++ bin/download_converted_resources.py
@@ -5,7 +5,7 @@ import click
 from datetime import datetime
 
 from bin.resources import get_resources
-from file_downloader import download_urls
+from bin.file_downloader import download_urls
 
 logger =  logging.getLogger("__name__")
